chg: Migrate to testinfra-bdd.
