+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
     "setuptools>=42",
-    "urllib3>=1.26.0",
+    "urllib3>=1.26.0,<2.0.0",
     "wheel"
 ]
 build-backend = "setuptools.build_meta"
