+++ requirements.txt
@@ -1,101 +1,39 @@
-alabaster==0.7.12
-attrs==21.2.0
-awscli==1.22.24
-awscli-local==0.17
-Babel==2.9.1
-bcrypt==3.2.0
-bleach==4.1.0
-boto3==1.20.24
-botocore==1.23.24
-build==0.7.0
-certifi==2021.10.8
-cffi==1.14.6
-charset-normalizer==2.0.6
-colorama==0.4.3
-coverage==6.0.1
-cryptography==35.0.0
-distro==1.6.0
-docker==5.0.3
-docker-compose==1.29.2
-dockerpty==0.4.1
-docopt==0.6.2
-docutils==0.15.2
-flake8==3.9.2
-flake8-docstrings==1.6.0
-flake8-quotes==3.3.0
+attrs==21.4.0
+bandit==1.7.4
+colorama==0.4.5
+flake8==4.0.1
 future==0.18.2
-gitchangelog==3.0.4
+gitdb==4.0.9
+GitPython==3.1.27
 glob2==0.7
-html2text==2020.1.16
-idna==3.2
-imagesize==1.2.0
-importlib-metadata==4.8.1
 iniconfig==1.1.1
-Jinja2==3.0.2
-jmespath==0.10.0
-jsonschema==3.2.0
-keyring==23.2.1
-localstack-client==1.28
-Mako==1.1.5
+Mako==1.2.1
 mando==0.6.4
-MarkupSafe==2.0.1
+MarkupSafe==2.1.1
 mccabe==0.6.1
-packaging==21.0
-paramiko==2.8.0
+numpy==1.22.3
+ofxtools==0.9.5
+packaging==21.3
+pandas==1.4.2
 parse==1.19.0
-parse-type==0.5.2
-pathspec==0.9.0
-pep517==0.11.0
-pkginfo==1.7.1
+parse-type==0.6.0
+pbr==5.9.0
 pluggy==1.0.0
-py==1.10.0
-pyasn1==0.4.8
-pycodestyle==2.7.0
-pycparser==2.20
-pydash==5.1.0
+py==1.11.0
+pycodestyle==2.8.0
 pydocstyle==6.1.1
-pyflakes==2.3.1
-Pygments==2.10.0
-PyNaCl==1.4.0
-pyparsing==2.4.7
-pyrsistent==0.18.0
-pystache==0.5.4
-pytest==6.2.5
-pytest-bdd==4.1.0
-pytest-cov==3.0.0
-pytest-flake8==1.0.7
-pytest-testinfra==6.4.0
+pyflakes==2.4.0
+pyparsing==3.0.9
+pytest==7.1.2
+pytest-bdd==6.0.1
+pytest-testinfra==6.8.0
 python-dateutil==2.8.2
-python-dotenv==0.19.1
-pytz==2021.3
-PyYAML==5.4.1
+pytz==2022.1
+PyYAML==6.0
 radon==5.1.0
-readme-renderer==30.0
-requests==2.26.0
-requests-toolbelt==0.9.1
-rfc3986==1.5.0
-rsa==4.7.2
-s3transfer==0.5.0
 six==1.16.0
-snowballstemmer==2.1.0
-Sphinx==4.2.0
-sphinx-markdown-builder==0.5.4
-sphinxcontrib-applehelp==1.0.2
-sphinxcontrib-devhelp==1.0.2
-sphinxcontrib-htmlhelp==2.0.0
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.3
-sphinxcontrib-serializinghtml==1.1.5
-texttable==1.6.4
-toml==0.10.2
-tomli==1.2.1
-tqdm==4.62.3
-twine==3.4.2
-unify==0.5
-untokenize==0.1.1
-urllib3==1.26.7
-webencodings==0.5.1
-websocket-client==0.59.0
-yamllint==1.26.3
-yapf==0.31.0
-zipp==3.6.0
+smmap==5.0.0
+snowballstemmer==2.2.0
+stevedore==4.0.0
+testinfra-bdd==2.0.0
+tomli==2.0.1
