+++ setup.cfg
@@ -10,7 +10,7 @@ source =
 docstring-convention: numpy
 max-line-length = 120
 per-file-ignores =
-    tests/*:D104,D401,D403
+    tests/*:D100,D104,D401,D403
 radon-max-cc: 5
 
 [tool:pytest]
