+++ tests/features/wait4localstack.feature
@@ -18,16 +18,3 @@ Feature: Wait4Localstack
       | cli_options       | return_value |
       | --verbose         | 0            |
       | --debug --verbose | 1            |
-
-  Scenario Outline: TestInfra
-    Given sut
-    When package is <testinfra_package>
-    And resource is <resource_name>
-    Then expect package response to match <package_response>
-    Examples:
-      | testinfra_package | resource_name                             | package_response |
-      | file.exists       | /tmp/dist/wait4localstack-$VERSION.tar.gz | True             |
-      | file.exists       | /usr/local/bin/entrypoint.sh              | True             |
-      | file.exists       | /usr/local/bin/wait4localstack            | True             |
-      | host.run          | /usr/local/bin/wait4localstack -h         | 0                |
-      | pip.is_installed  | wait4localstack                           | True             |
